+++ src/frontend/js/pages/TeacherDashboardContractsLoader/TeacherDashboardOrganizationContractsLoader/index.tsx
@@ -1,10 +1,7 @@
-import { FormattedMessage, defineMessages } from 'react-intl';
-import { useParams } from 'react-router-dom';
-import queryString from 'query-string';
+import { defineMessages, FormattedMessage } from 'react-intl';
 
 import { DashboardLayout } from 'widgets/Dashboard/components/DashboardLayout';
 import { TeacherDashboardOrganizationSidebar } from 'widgets/Dashboard/components/TeacherDashboardOrganizationSidebar';
-import { location } from 'utils/indirection/window';
 import TeacherDashboardContracts from '../TeacherDashboardContracts';
 
 const messages = defineMessages({
@@ -13,16 +10,9 @@ const messages = defineMessages({
     description: 'Use for the page title of the organization contracts area',
     id: 'pages.TeacherDashboardOrganizationContractsLoader.pageTitle',
   },
-  loading: {
-    defaultMessage: 'Loading contracts...',
-    description: 'Message displayed while loading contracts',
-    id: 'pages.TeacherDashboardOrganizationContractsLoader.loading',
-  },
 });
 
 export const TeacherDashboardOrganizationContractsLoader = () => {
-  const { organizationId } = useParams<{ organizationId?: string }>();
-  const { page = '1' }: { page?: string } = queryString.parse(location.search);
   return (
     <DashboardLayout sidebar={<TeacherDashboardOrganizationSidebar />}>
       <div className="dashboard__page_title_container">
@@ -30,7 +20,7 @@ export const TeacherDashboardOrganizationContractsLoader = () => {
           <FormattedMessage {...messages.pageTitle} />
         </h1>
       </div>
-      <TeacherDashboardContracts organizationId={organizationId} page={page} />
+      <TeacherDashboardContracts />
     </DashboardLayout>
   );
 };